잔동충전 > coins 맴버변수의 위치를 모델로 변경한다
- 추가로 데이터를 가공하고 저장하는 로직을 model로 이동한다
